requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-telegram-bot==21.3
requests==2.31.0
waitress==2.1.2
flask==3.0.3
APScheduler==3.10.4
pytz==2024.1
